Made MCU wait until GPS threshold voltage is reached.
Made it a priority for GPS to get a fix before starting to do anything else(e.g.  send Radio join requests). I have set the GPS to search for a fix FOREVER and cannot do anything until it gets a fix. Realistically, if we don't know where we are, we won't know what LoRa params to set.