put in code to save the APP_KEY, JOIN EUI, NWK_KEY, DEVICE_EUI to EEPROM
IN the code, these values are never saved to EEPROM, but I have implemented the functions to do so, taken from the code for the STEVAL-STRKT01 tracker. The code for the STRKT01 can be found here: https://www.st.com/en/embedded-software/fp-atr-lora1.html

It appears that the code on the STRKT01 is an evolution of the code in the I-LORAWAN-CUBE software stack, which is the base for the pico tracker lora code.

I think there is no real need to save the keys to EEPROM.