Ensure counter increments (#65)
* make seperate function to increment eeprom indexes

* puti n seperate function to playback_hw_init()

* put in seperate function to update reset counts

* do iwdg refresh whenever reading from EEPROM

* add function to pretty print sensor values

* put seperate function to save data to nvm

* update documentation

* put in function to fill structs to send to gnd

* update comment

* formatting and adding comments

* print normally

* [needs improvement] put in workaround to increment index counter

In the very first transmission, it skips writing to the first index. Then later, the program reads this first index, all 0s, and gets confused. Workaround is not to skip on the very first time.

* fix casting bug

* set option to use watchdog

* memset gps info to all zeros at start of gps search

Necessary to redefine GPS_FAILURE = 0 because when memsetting all to zero, I want to ensure the default is failure.

* calculate days of playback based on most recent saved position

Dont base it on GPS fix, incase the gps fix is not successful. Otherwise, it would calculate gps time on the default value e.g. 00:00:00h

* formatting only

* print out debug info

Co-authored-by: MedadRufus <mailto@medadnewman.co.uk>